pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lunar"
version = "0.1.0"
dependencies = [
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "fonttools==4.60.1",
    "kiwisolver==1.4.9",
    "matplotlib==3.10.7",
    "neat-python==0.92",
    "noise==1.2.2",
    "numpy==2.3.4",
    "packaging==25.0",
    "pillow==12.0.0",
    "pyparsing==3.2.5",
    "python-dateutil==2.9.0.post0",
    "pyyaml==6.0.3",
    "six==1.17.0",
]
description = "LUNAR simulation project"
requires-python = "==3.11.9"

[tool.setuptools.packages.find]
where = ["src"]
include = ["python_sim*"]
